Fix CodeQL warnings
Resolved code scanning alerts:

java/xxe (critical) - harden DocumentBuilderFactory against XXE:
- https://github.com/BearWare/TeamTalk5/security/code-scanning/47
- https://github.com/BearWare/TeamTalk5/security/code-scanning/48
- https://github.com/BearWare/TeamTalk5/security/code-scanning/49
- https://github.com/BearWare/TeamTalk5/security/code-scanning/50

java/path-injection (high) - sanitize untrusted display name before building cache file:
- https://github.com/BearWare/TeamTalk5/security/code-scanning/54
- https://github.com/BearWare/TeamTalk5/security/code-scanning/55
- https://github.com/BearWare/TeamTalk5/security/code-scanning/56
- https://github.com/BearWare/TeamTalk5/security/code-scanning/57
- https://github.com/BearWare/TeamTalk5/security/code-scanning/58
- https://github.com/BearWare/TeamTalk5/security/code-scanning/59

cpp/suspicious-add-sizeof (high) - use element count instead of sizeof for array end:
- https://github.com/BearWare/TeamTalk5/security/code-scanning/881

cpp/comparison-with-wider-type (high) - widen loop counter to size_t:
- https://github.com/BearWare/TeamTalk5/security/code-scanning/1

cpp/integer-multiplication-cast-to-long (high) - cast operand to size_t before multiply:
- https://github.com/BearWare/TeamTalk5/security/code-scanning/4
- https://github.com/BearWare/TeamTalk5/security/code-scanning/5
- https://github.com/BearWare/TeamTalk5/security/code-scanning/8
- https://github.com/BearWare/TeamTalk5/security/code-scanning/14
- https://github.com/BearWare/TeamTalk5/security/code-scanning/21
- https://github.com/BearWare/TeamTalk5/security/code-scanning/23
- https://github.com/BearWare/TeamTalk5/security/code-scanning/24
- https://github.com/BearWare/TeamTalk5/security/code-scanning/34
- https://github.com/BearWare/TeamTalk5/security/code-scanning/35
- https://github.com/BearWare/TeamTalk5/security/code-scanning/36
- https://github.com/BearWare/TeamTalk5/security/code-scanning/37
- https://github.com/BearWare/TeamTalk5/security/code-scanning/38
- https://github.com/BearWare/TeamTalk5/security/code-scanning/39
- https://github.com/BearWare/TeamTalk5/security/code-scanning/41
- https://github.com/BearWare/TeamTalk5/security/code-scanning/42
- https://github.com/BearWare/TeamTalk5/security/code-scanning/43
- https://github.com/BearWare/TeamTalk5/security/code-scanning/44
- https://github.com/BearWare/TeamTalk5/security/code-scanning/45
- https://github.com/BearWare/TeamTalk5/security/code-scanning/46
- https://github.com/BearWare/TeamTalk5/security/code-scanning/71
- https://github.com/BearWare/TeamTalk5/security/code-scanning/72
- https://github.com/BearWare/TeamTalk5/security/code-scanning/73
- https://github.com/BearWare/TeamTalk5/security/code-scanning/74
- https://github.com/BearWare/TeamTalk5/security/code-scanning/75
- https://github.com/BearWare/TeamTalk5/security/code-scanning/76
- https://github.com/BearWare/TeamTalk5/security/code-scanning/77
- https://github.com/BearWare/TeamTalk5/security/code-scanning/78
- https://github.com/BearWare/TeamTalk5/security/code-scanning/79
- https://github.com/BearWare/TeamTalk5/security/code-scanning/80
- https://github.com/BearWare/TeamTalk5/security/code-scanning/81
- https://github.com/BearWare/TeamTalk5/security/code-scanning/82
- https://github.com/BearWare/TeamTalk5/security/code-scanning/83
- https://github.com/BearWare/TeamTalk5/security/code-scanning/84
- https://github.com/BearWare/TeamTalk5/security/code-scanning/870
- https://github.com/BearWare/TeamTalk5/security/code-scanning/871

+++ Library/TeamTalkLib/avstream/AudioInputStreamer.cpp
@@ -162,8 +162,8 @@ bool AudioInputStreamer::ProcessResample()
         assert(frame.inputfmt == m_inputfmt);
         assert(m_resampler);
         int const osamples = CalcSamples(frame.inputfmt.samplerate, frame.input_samples, GetMediaOutput().audio.samplerate);
-        if (m_resamplebuffer.size() != osamples * GetMediaOutput().audio.channels)
-            m_resamplebuffer.resize(osamples * GetMediaOutput().audio.channels);
+        if (m_resamplebuffer.size() != size_t(osamples) * GetMediaOutput().audio.channels)
+            m_resamplebuffer.resize(size_t(osamples) * GetMediaOutput().audio.channels);
         ret = m_resampler->Resample(frame.input_buffer, frame.input_samples, m_resamplebuffer.data(), osamples);
         assert(ret > 0);
         media::AudioFrame const resam_frame(GetMediaOutput().audio, m_resamplebuffer.data(), osamples);
